Moved chart to component
